refactored getPriority() to return an int for reverse compatibility   changed values of priorities to make zero default   made throw an IllegalStateException if the priority is out of range fixed up regression tests:   exit with proper error code   better operation for simulator   scribe regression test works with consistency algorithm   tested past and scribe   splitstream and replication fail in simulator fixed logging/CME/and a bug in ScribeImpl   the bug: very old bug, when copying from first impl, was randomly unsubscribing a leafset neighbor fixed logging in BasicNetworkSimulator changed logging levels to INFO for many common WARNINGs SelectorManager uses a PriorityQueue rather than the TreeSet -- better performance
git-svn-id: https://svn.mpi-sws.org/DS/freepastry/trunk@3477 c21dacaf-8602-0410-aacd-8bc41df2cbe7

+++ pastry/src/rice/p2p/aggregation/AggregationImpl.java
@@ -168,7 +168,7 @@ public class AggregationImpl implements GCPast, VersioningPast, Aggregation, App
 
     this.endpoint = node.buildEndpoint(this, instance);
     this.endpoint.setDeserializer(new MessageDeserializer() {    
-      public Message deserialize(InputBuffer buf, short type, byte priority,
+      public Message deserialize(InputBuffer buf, short type, int priority,
           NodeHandle sender) throws IOException {
         switch(type) {
 //          case AggregationTimeoutMessage.TYPE:
