+++ pastry/src/rice/p2p/glacier/v2/GlacierImpl.java
@@ -233,7 +233,7 @@ public class GlacierImpl implements Glacier, GCPast, VersioningPast, Application
     this.endpoint = node.buildEndpoint(this, instance);
     this.endpoint.setDeserializer(new MessageDeserializer() {
     
-      public Message deserialize(InputBuffer buf, short type, byte priority,
+      public Message deserialize(InputBuffer buf, short type, int priority,
           NodeHandle sender) throws IOException {
         switch(type) {
           case GlacierDataMessage.TYPE:
