+++ pastry/src/rice/p2p/util/rawserialization/JavaSerializedDeserializer.java
@@ -29,7 +29,7 @@ public class JavaSerializedDeserializer implements MessageDeserializer {
     deserializeOnlyTypeZero = !val; 
   }
   
-  public Message deserialize(InputBuffer buf, short type, byte priority, NodeHandle sender) throws IOException {
+  public Message deserialize(InputBuffer buf, short type, int priority, NodeHandle sender) throws IOException {
     if (deserializeOnlyTypeZero && (type != 0)) throw new IllegalArgumentException("Type must be zero, was "+type+".  See http://freepastry.org/FreePastry/extendingRawMessages.html for more information.");
