+++ pastry/src/rice/post/messaging/PostPastryMessage.java
@@ -36,7 +36,7 @@ public class PostPastryMessage implements RawMessage, SignedPostMessageWrapper {
    *
    * @return This message's priority
    */
-  public byte getPriority() {
+  public int getPriority() {
     return MEDIUM_HIGH_PRIORITY;
   }
